requirements.txt
# This project requires Python 3.10 or higher

# Core data processing and analysis
numpy==1.26.0
pandas==2.0.0
plotly==5.18.0
scipy==1.10.1
pandas_ta==0.3.14b0

# Database and ORM
SQLAlchemy==2.0.23

# API and networking
python-dotenv==1.0.0
yfinance==0.2.31
requests==2.31.0

# Testing
pytest==7.4.3
psutil==6.0.0

# Trading and backtesting
lumibot==2.9.0
mplfinance==0.12.7a17

# Development tools
black==23.11.0
mypy==1.7.0
isort==5.12.0
jupyter==1.0.0

# Frontend
streamlit==1.29.0